requirements.txt
annotated-types==0.7.0
gitdb==4.0.12
GitPython==3.1.41
markdown-it-py==3.0.0
mdurl==0.1.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
rich==14.0.0
six==1.17.0
smmap==5.0.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1